# hw2/src/main/resources/openapi/marketplace.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  version: 1.0.0
servers:
  - url: http://localhost:8080
tags:
  - name: Auth
  - name: Products
  - name: Orders
  - name: PromoCodes
paths:
  /auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: Регистрация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Пользователь зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Аутентификация
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Токены выданы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refresh
      summary: Обновление access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Токены обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /products:
    get:
      tags: [Products]
      operationId: listProducts
      summary: Список товаров
      description: Доступно ролям USER, SELLER и ADMIN.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            default: 20
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProductStatus'
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Страница товаров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Products]
      operationId: createProduct
      summary: Создание товара
      description: SELLER создаёт свой товар, ADMIN может создать любой товар.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Товар создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /products/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Products]
      operationId: getProduct
      summary: Получение товара
      description: Доступно ролям USER, SELLER и ADMIN.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Товар
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Products]
      operationId: updateProduct
      summary: Обновление товара
      description: SELLER обновляет только свои товары, ADMIN — любые.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Товар обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Products]
      operationId: deleteProduct
      summary: Архивация товара
      description: SELLER архивирует только свои товары, ADMIN — любые.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Товар переведён в ARCHIVED
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /orders:
    post:
      tags: [Orders]
      operationId: createOrder
      summary: Создание заказа
      description: Доступно ролям USER и ADMIN.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Заказ создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /orders/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Orders]
      operationId: getOrder
      summary: Получение заказа
      description: USER получает только свой заказ, ADMIN — любой. SELLER доступ запрещён.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Заказ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Orders]
      operationId: updateOrder
      summary: Обновление заказа
      description: USER обновляет только свой заказ, ADMIN — любой. SELLER доступ запрещён.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdate'
      responses:
        '200':
          description: Заказ обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /orders/{id}/cancel:
    post:
      tags: [Orders]
      operationId: cancelOrder
      summary: Отмена заказа
      description: USER отменяет только свой заказ, ADMIN — любой. SELLER доступ запрещён.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Заказ отменён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /promo-codes:
    post:
      tags: [PromoCodes]
      operationId: createPromoCode
      summary: Создание промокода
      description: Доступно ролям SELLER и ADMIN.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromoCodeCreate'
      responses:
        '201':
          description: Промокод создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromoCodeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: VALIDATION_ERROR
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: TOKEN_EXPIRED, TOKEN_INVALID или REFRESH_TOKEN_INVALID
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: ACCESS_DENIED или ORDER_OWNERSHIP_VIOLATION
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: PRODUCT_NOT_FOUND или ORDER_NOT_FOUND
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: PRODUCT_INACTIVE, ORDER_HAS_ACTIVE, INVALID_STATE_TRANSITION или INSUFFICIENT_STOCK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    UnprocessableEntity:
      description: PROMO_CODE_INVALID или PROMO_CODE_MIN_AMOUNT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    TooManyRequests:
      description: ORDER_LIMIT_EXCEEDED
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    Role:
      type: string
      enum: [USER, SELLER, ADMIN]
    ProductStatus:
      type: string
      enum: [ACTIVE, INACTIVE, ARCHIVED]
    OrderStatus:
      type: string
      enum: [CREATED, PAYMENT_PENDING, PAID, SHIPPED, COMPLETED, CANCELED]
    DiscountType:
      type: string
      enum: [PERCENTAGE, FIXED_AMOUNT]
    RegisterRequest:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
        password:
          type: string
          minLength: 8
          maxLength: 72
        role:
          $ref: '#/components/schemas/Role'
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          minLength: 1
    AuthResponse:
      type: object
      required: [access_token, refresh_token]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
    ProductCreate:
      type: object
      required: [name, price, stock, category, status]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 4000
          nullable: true
        price:
          type: number
          format: decimal
          minimum: 0.01
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
          maxLength: 100
        status:
          $ref: '#/components/schemas/ProductStatus'
        seller_id:
          type: string
          format: uuid
          nullable: true
          description: ADMIN может указать продавца; для SELLER используется его id.
    ProductUpdate:
      type: object
      required: [name, price, stock, category, status]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 4000
          nullable: true
        price:
          type: number
          format: decimal
          minimum: 0.01
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
          maxLength: 100
        status:
          $ref: '#/components/schemas/ProductStatus'
    ProductResponse:
      type: object
      required: [id, name, price, stock, category, status, seller_id, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: decimal
        stock:
          type: integer
        category:
          type: string
        status:
          $ref: '#/components/schemas/ProductStatus'
        seller_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProductPage:
      type: object
      required: [items, totalElements, page, size]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        totalElements:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer
    OrderItemRequest:
      type: object
      required: [product_id, quantity]
      properties:
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
          maximum: 999
    OrderCreate:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/OrderItemRequest'
        promo_code:
          type: string
          pattern: '^[A-Z0-9_]{4,20}$'
          nullable: true
    OrderUpdate:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/OrderItemRequest'
    OrderItemResponse:
      type: object
      required: [product_id, quantity, price_at_order]
      properties:
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
        price_at_order:
          type: number
          format: decimal
    OrderResponse:
      type: object
      required: [id, user_id, status, items, total_amount, discount_amount, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/OrderStatus'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        promo_code:
          type: string
          nullable: true
        total_amount:
          type: number
          format: decimal
        discount_amount:
          type: number
          format: decimal
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PromoCodeCreate:
      type: object
      required: [code, discount_type, discount_value, min_order_amount, max_uses, valid_from, valid_until]
      properties:
        code:
          type: string
          pattern: '^[A-Z0-9_]{4,20}$'
        discount_type:
          $ref: '#/components/schemas/DiscountType'
        discount_value:
          type: number
          format: decimal
          minimum: 0.01
        min_order_amount:
          type: number
          format: decimal
          minimum: 0
        max_uses:
          type: integer
          minimum: 1
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        active:
          type: boolean
          default: true
    PromoCodeResponse:
      type: object
      required: [id, code, discount_type, discount_value, min_order_amount, max_uses, current_uses, valid_from, valid_until, active]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        discount_type:
          $ref: '#/components/schemas/DiscountType'
        discount_value:
          type: number
          format: decimal
        min_order_amount:
          type: number
          format: decimal
        max_uses:
          type: integer
        current_uses:
          type: integer
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        active:
          type: boolean
    ApiError:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          type: string
          enum:
            - PRODUCT_NOT_FOUND
            - PRODUCT_INACTIVE
            - ORDER_NOT_FOUND
            - ORDER_LIMIT_EXCEEDED
            - ORDER_HAS_ACTIVE
            - INVALID_STATE_TRANSITION
            - INSUFFICIENT_STOCK
            - PROMO_CODE_INVALID
            - PROMO_CODE_MIN_AMOUNT
            - ORDER_OWNERSHIP_VIOLATION
            - VALIDATION_ERROR
            - TOKEN_EXPIRED
            - TOKEN_INVALID
            - REFRESH_TOKEN_INVALID
            - ACCESS_DENIED
        message:
          type: string
        details:
          type: object
          nullable: true
          additionalProperties: true